backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.18
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.40.0
open-clip-torch>=2.29.0
faiss-cpu>=1.9.0
umap-learn>=0.5.7
hdbscan>=0.8.40
scikit-learn>=1.6.0
numpy>=1.26.0
Pillow>=11.0.0
supabase>=2.13.0
httpx>=0.28.0
beautifulsoup4>=4.13.0
pinterest-export>=0.1.0
playwright>=1.40.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.0
datasets>=3.0.0
